requirements.txt
# 核心依賴套件 (深度學習框架)
ultralytics>=8.0.0,<9.0.0
torch>=1.13.0,<3.0.0
torchvision>=0.14.0,<1.0.0

# 數值計算和圖像處理
numpy>=1.21.0,<2.0.0
opencv-python>=4.5.0,<5.0.0
Pillow>=8.0.0,<11.0.0

# 配置和數據處理
PyYAML>=6.0,<7.0.0
matplotlib>=3.5.0,<4.0.0
tqdm>=4.64.0,<5.0.0
pandas>=1.5.0,<3.0.0

# 超參數優化
optuna>=3.0.0,<4.0.0
plotly>=5.0.0,<6.0.0

# 系統監控
psutil>=5.9.0,<6.0.0
GPUtil>=1.4.0,<2.0.0

# 日誌和可視化 (可選)
tensorboard>=2.10.0,<3.0.0
wandb>=0.13.0,<1.0.0

# 工具套件
typing-extensions>=4.0.0,<5.0.0

# 開發工具 (可選)
# jupyter>=1.0.0,<2.0.0  # Jupyter 環境
# pytest>=7.0.0,<8.0.0   # 測試框架

# 可選依賴（根據環境選擇安裝）
# jupyter>=1.0.0  # Jupyter 環境
# google-colab     # Colab 環境
